emoji-favico/requirements.txt
python-fasthtml
uvicorn
requests
jupyter
emoji
